refactor(string): split 477-line module into string-transform, string-query, string-template, string-validators
src/string.ts is now a thin barrel re-exporting from four focused sub-modules
under src/_internal/:

- string-transform.ts   — toUpperCase, toLowerCase, trim, trimStart, trimEnd,
                          split, join, replace, replaceAll, capitalize, camelCase,
                          pascalCase, snakeCase, kebabCase, titleCase, slice,
                          substring, padStart, padEnd, repeat, reverse, truncate,
                          words, lines, indent, dedent
- string-query.ts       — startsWith, endsWith, includes, matches, isEmptyString,
                          isBlank, length
- string-template.ts    — template, format
- string-validators.ts  — isEmail, isUrl, isNumeric, isAlpha, isAlphanumeric,
                          isHexColor, isUUID

All public exports and the fp-core/string subpath import remain unchanged.

Closes #28
